+++ src/test/java/ru/gigastack/lineProcessing/impl/LineClassifierImplTest.java
@@ -5,11 +5,12 @@ import ru.gigastack.model.DataType;
 import ru.gigastack.lineProcessing.TypeParser;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class LineClassifierImplTest {
-    private final TypeParser<Long> intParser = new IntegerParser();
+    private final TypeParser<BigInteger> intParser = new IntegerParser();
     private final TypeParser<BigDecimal> floatParser = new FloatParser();
     private final LineClassifierImpl classifier = new LineClassifierImpl(intParser, floatParser);
